users_service/requirements.txt
django
djangorestframework
gunicorn
requests
djangorestframework-simplejwt